+++ NEWS.md
@@ -1,3 +1,6 @@
+## gDRcore 1.1.20 - 2024-05-16
+* move `define_matrix_grid_positions` function to `gDRutils` package
+
 ## gDRcore 1.1.20 - 2024-05-08
 * fix typo